fix(performance): remove all uses of memoizee
While there may be performance problems from checking the state of the
file system, we need to find a way to resolve that which *does not*
include memoizing calls for anywhere from 1 to 10 minutes. We may
reintroduce some kind of caching later, but for now we need to simply
start by iterating on the architecture so that it does less work,
especially for large apps, where this will cause extremely uneven and
inconsistent performance whenever the *non*-memoized version of the
call is invoked.