chore(dependencies): bump TypeScript to 3.6
Updating the compiler version resolves some major performance problems
triggered by the Glimmer type definitions. Unfortunately, it also causes
some noise due to [glimmerjs/glimmer-vm#946], but that should be fixed
in an upcoming release of the `@glimmer/*` packages, and it does not
cause internal issues for this package.

[glimmerjs/glimmer-vm#946]: https://github.com/glimmerjs/glimmer-vm/issues/946

Also update the tsconfig to target current versions of JavaScript in
build and in lib to use.